Playability improved. Boost lasts twice as long, level up in half the time.

+++ GameCode/Race/RaceControls/RacingControls.cs
@@ -354,7 +354,7 @@ namespace BeatShift.Input
             return (lastTaps[tapNo]);
         }
 
-        public int getBoostValue()
+        public double getBoostValue()
         {
             return racer.beatQueue.GetBoost();
         }
